Added all the repo and sone services files are defined

+++ src/main/java/com/application/Object/address.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 public class address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int address_id;
+    private Long address_id;
     private String street;
     private String city;
     private String pincode;
